+++ src/adaptors/wombat-exchange/config.js
@@ -14,10 +14,4 @@ module.exports = {
     WOM_ADDRESS: '0x7b5eb3940021ec0e8e463d5dbb4b7b09a89ddf96',
   },
 
-  avax:{
-    rpcUrl: 'https://api.avax.network/ext/bc/C/rpc',
-    ggAVAXRewarder:'0x3Bc2Fe0f365c8559510419b8b10e43e275D8aDC8',
-    AVAXRewarder: '0x5EF81b1b2F4629f99E6724475003Be3796059638',
-  },
-
 };
